MemUtils: breakup into platform specific files
