+++ xbmc/platform/android/CMakeLists.txt
@@ -1,0 +1,3 @@
+set(SOURCES MemUtils.cpp)
+
+core_add_library(androidsupport)
